requirements.txt
BeautifulSoup==3.2.1
Django==1.7.4
Markdown==2.2.1
MySQL-python==1.2.3
PIL==1.1.7
argparse==1.2.1
cssselect==0.7.1
distribute==0.6.24
django-appconf==0.6
django-compressor==1.4
django-filebrowser==3.5.7
django-grappelli==2.5.6
django-mptt==0.6.1
django-taggit==0.12.2
lxml>=2.3.4
pyquery==1.2.1
versiontools==1.9.1
wsgiref==0.1.2
gunicorn
psycopg2
dj-database-url
